Add Dependabot for go
Happened to notice Dependabot isn't setup for go modules. 

When I checked, I realized that is because the `go.mod` file has no external dependencies.

However, there's always a chance that could change in the future, and if so we'd probably forget to update here...
So thought it doesn't hurt to preemptively add it here.